+++ angular-workspace/projects/ni/nimble-angular/src/directives/combobox/tests/nimble-combobox.directive.spec.ts
@@ -1,9 +1,11 @@
 import { Component, ElementRef, ViewChild } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import type { BooleanValueOrAttribute } from '@ni/nimble-angular/internal-utilities';
-import { ComboboxAutocomplete, DropdownAppearance } from '../../../public-api';
+import { FormsModule } from '@angular/forms';
+import { ComboboxAutocomplete, DropdownAppearance, NimbleListOptionModule } from '../../../public-api';
 import { Combobox, NimbleComboboxDirective } from '../nimble-combobox.directive';
 import { NimbleComboboxModule } from '../nimble-combobox.module';
+import { waitForUpdatesAsync } from '../../../testing/async-helpers';
 
 describe('Nimble combobox', () => {
     beforeEach(() => {
@@ -356,6 +358,89 @@ describe('Nimble combobox', () => {
 
             expect(directive.errorVisible).toBeTrue();
             expect(nativeElement.errorVisible).toBeTrue();
+        });
+    });
+
+    describe('can access value through directive', () => {
+        interface TestModel {
+            name: string;
+            value: number;
+        }
+
+        @Component({
+            template: `
+                <nimble-combobox #combobox>
+                    <nimble-list-option *ngFor="let option of autoCompleteOptions" [ngValue]="option">{{ option.name }}</nimble-list-option>
+                </nimble-combobox>
+            `
+        })
+        class TestHostComponent {
+            @ViewChild('combobox', { read: NimbleComboboxDirective }) public directive: NimbleComboboxDirective;
+            @ViewChild('combobox', { read: ElementRef }) public elementRef: ElementRef<Combobox>;
+            public autoCompleteOptions: TestModel[] = [
+                { name: 'Zero', value: 0 },
+                { name: 'One', value: 1 },
+                { name: 'Two', value: 2 }
+            ];
+        }
+
+        let fixture: ComponentFixture<TestHostComponent>;
+        let directive: NimbleComboboxDirective;
+        let nativeElement: Combobox;
+
+        function typeValueInCombobox(value: string): void {
+            nativeElement.control.value = value;
+            const inputEvent = new InputEvent('input', { data: value, inputType: 'insertText' });
+            nativeElement.inputHandler(inputEvent);
+            nativeElement.dispatchEvent(inputEvent);
+        }
+
+        function clickComboboxOption(optionIndex: number): void {
+            nativeElement.dispatchEvent(new Event('click'));
+            nativeElement.options[optionIndex].dispatchEvent(new Event('click', { bubbles: true }));
+        }
+
+        beforeEach(async () => {
+            TestBed.configureTestingModule({
+                declarations: [TestHostComponent],
+                imports: [NimbleComboboxModule, NimbleListOptionModule, FormsModule]
+            });
+            fixture = TestBed.createComponent(TestHostComponent);
+            fixture.detectChanges();
+            // wait for combobox's 'options' property to be updated from slotted content
+            await waitForUpdatesAsync();
+            directive = fixture.componentInstance.directive;
+            nativeElement = fixture.componentInstance.elementRef.nativeElement;
+        });
+
+        it('before a value is set', () => {
+            expect(directive.value).toBe('');
+            expect(nativeElement.value).toBe('');
+        });
+
+        it('after typing a value not in the auto complete list', () => {
+            const typedText = 'hello world';
+            typeValueInCombobox(typedText);
+
+            expect(directive.value).toBe(typedText);
+            expect(nativeElement.value).toBe(typedText);
+        });
+
+        it('after typing a value in the auto complete list', () => {
+            const typedText = fixture.componentInstance.autoCompleteOptions[0].name;
+            typeValueInCombobox(typedText);
+
+            expect(directive.value).toBe(typedText);
+            expect(nativeElement.value).toBe(typedText);
+        });
+
+        it('after clicking a value in the auto complete list', () => {
+            const optionClickIndex = 1;
+            const option = fixture.componentInstance.autoCompleteOptions[optionClickIndex];
+            clickComboboxOption(optionClickIndex);
+
+            expect(directive.value).toBe(option.name);
+            expect(nativeElement.value).toBe(option.name);
         });
     });
 });