+++ packages/nimble-components/src/table-column/base/types.ts
@@ -12,8 +12,8 @@ export interface TableCellState<
     TCellRecord extends TableCellRecord = TableCellRecord,
     TColumnConfig = unknown
 > {
-    cellRecord: TCellRecord;
-    columnConfig: TColumnConfig;
+    cellRecord?: TCellRecord;
+    columnConfig?: TColumnConfig;
 }
 
 /**