+++ packages/nimble-components/src/table-column/date-text/index.ts
@@ -31,6 +31,7 @@ import type {
 } from './types';
 import { TableColumnDateTextValidator } from './models/table-column-date-text-validator';
 import { lang } from '../../theme-provider';
+import { optionalBooleanConverter } from '../../utilities/models/converter';
 
 export type TableColumnDateTextCellRecord = TableNumberField<'value'>;
 export interface TableColumnDateTextColumnConfig {
@@ -86,7 +87,10 @@ export class TableColumnDateText extends TableColumnTextBase {
     @attr({ attribute: 'custom-format-matcher' })
     public customFormatMatcher: FormatMatcherAlgorithm;
 
-    @attr({ attribute: 'custom-hour12', mode: 'boolean' })
+    // Later versions of FAST (than the legacy branch we're on) have a nullableBooleanConverter.
+    // We should replace our converter with that one when it is available to us.
+    // See issue related to adopting FastElement 2.0: https://github.com/ni/nimble/issues/572
+    @attr({ attribute: 'custom-hour12', converter: optionalBooleanConverter })
     public customHour12?: boolean;
 
     @attr({ attribute: 'custom-time-zone' })